refactor: extract large inline test modules
